avsted_requirements.txt
python-dotenv==1.0.1
langchain
langchain-openai
boto3
langgraph
#grandalf==0.8
sendgrid==6.11.0
streamlit==1.38.0
serpapi==0.1.5